+++ controllers/api/allergen-routes.js
@@ -5,13 +5,11 @@ const { Allergen } = require('../../models');
 router.post('/', async (req, res) => {
     try {
       var user_id = 0;
-      console.log("first : ", req.body.user_id);
       if (!req.body.user_id) {
         user_id = req.session.user;
       } else {
         user_id = req.body.user_id;
       }
-      console.log("second : ", user_id);
       const dbAllergenData = await Allergen.create({
         allergen_name: req.body.allergen_name,
         user_id: user_id,
