+++ controllers/home-routes.js
@@ -1,5 +1,9 @@
 const router = require("express").Router();
 
+const fetch = require("node-fetch");
+const sequelize = require('../config/connection');
+require('dotenv').config();
+
 const { Allergen, Product } = require("../models");
 
 const withAuth = require("../utils/auth");
@@ -55,9 +59,44 @@ router.get("/login", async (req, res) => {
 router.get("/search", async (req, res) => {
   var searches = [];
   var allergens = [];
+  search_product = req.session.search_product;
+  search_allergen = req.session.search_allergen;
 
+	var apiUrl = `https://spoonacular-recipe-food-nutrition-v1.p.rapidapi.com/food/ingredients/autocomplete?query=${search_product}&number=5`;
+
+  if (search_allergen) {
+    apiUrl = apiUrl + `&intolerances=${search_allergen}`;
+  }
+  
+  //this only fetches name and images but you are able to filter by intolerances such as dairy, egg, gluten, peanut, sesame, seafood, shellfish, soy, sulfite, tree nut, and wheat.
+  await fetch(apiUrl, {
+    "method": "GET",
+    "headers": {
+      "x-rapidapi-host": "spoonacular-recipe-food-nutrition-v1.p.rapidapi.com",
+      "x-rapidapi-key": process.env.API_KEY
+    }
+  })
+  .then(response => response.json())
+  .then(data => 
+    {
+      for (let i = 0; i < data.length; i++){
+        var name = data[i].name;
+        var nameAttitute = name.replace(/\s/g, "%20");
+        var product_image = `https://spoonacular.com/cdn/ingredients_100x100/${data[i].image}`;
+        var searchItem = {
+          index: i,
+          name: name,
+          nameAttitute: nameAttitute,
+          product_image: product_image
+        };
+        searches.push(searchItem);
+      }
+    })
+  .catch(err => {
+    console.error(err);
+  })
+    
   if (req.session.loggedIn) {
-    searches = req.session.searchData;
 
     const allergenData = await Allergen.findAll({
       where: {
@@ -67,13 +106,24 @@ router.get("/search", async (req, res) => {
     allergens = allergenData.map((allergen) =>
       allergen.get({ plain: true })
     );
-  } 
+  }
 
   res.render("searchResult", {
     searches,
     allergens,
     loggedIn: req.session.loggedIn,
   });
+});
+
+router.post("/search", async (req, res) => {
+  req.session.search_product = req.body.search_product;
+  req.session.search_allergen = req.body.search_allergen;
+
+  const search_queries = {
+    product: req.session.search_product,
+    allergen: req.session.search_allergen,
+  }
+  res.status(200).json(search_queries);
 });
 
 module.exports = router;
