+++ package.json
@@ -19,7 +19,7 @@
     "express-session": "^1.17.1",
     "handlebars": "^4.7.6",
     "mysql2": "^2.3.3",
-    "node-fetch": "^3.1.0",
+    "node-fetch": "^2.6.1",
     "sequelize": "^6.12.4"
   }
 }
